Refactor VectorAlgebra
* Moved stuff in deference files.
* Generalized sum/prod/.. to 2 map
